+++ src/rchia/Blocks/BlockHeaderCommand.cs
@@ -14,14 +14,14 @@ internal sealed class BlockHeaderCommand : EndpointOptions
     public string? Hash { get; init; }
 
     [Option("h", "height", ArgumentHelpName = "HEIGHT", Description = "Look up a block by height")]
-    public uint? Height { get; init; }
+    public int? Height { get; init; }
 
     private async Task<string> GetHash(FullNodeProxy proxy)
     { 
         if (Height.HasValue)
         {
             using var cts = new CancellationTokenSource(TimeoutMilliseconds);
-            var block = await proxy.GetBlockRecordByHeight(Height.Value, cts.Token);
+            var block = await proxy.GetBlockRecordByHeight((uint)Height.Value, cts.Token);
 
             return block.HeaderHash;
         }
