+++ src/rchia/PlotNft/LeavePoolCommand.cs
@@ -8,7 +8,7 @@ namespace rchia.PlotNft;
 internal sealed class LeavePoolCommand : WalletCommand
 {
     [Option("i", "id", Default = 1, Description = "Id of the user wallet to use")]
-    public uint Id { get; init; } = 1;
+    public int Id { get; init; } = 1;
 
     [Option("f", "force", Description = "Do not prompt before nft creation")]
     public bool Force { get; init; }
@@ -26,7 +26,7 @@ internal sealed class LeavePoolCommand : WalletCommand
         return await DoWorkAsync("Leaving pool...", async output =>
         {
             using var rpcClient = await ClientFactory.Factory.CreateRpcClient(output, this, ServiceNames.Wallet);
-            var wallet = new PoolWallet(Id, await Login(rpcClient, output));
+            var wallet = new PoolWallet((uint)Id, await Login(rpcClient, output));
 
             using var cts = new CancellationTokenSource(TimeoutMilliseconds);
             await wallet.Validate(cts.Token);
