+++ src/rchia/Wallets/DeleteUnconfirmedTransactionsCommand.cs
@@ -8,7 +8,7 @@ namespace rchia.Wallet;
 internal sealed class DeleteUnconfirmedTransactionsCommand : WalletCommand
 {
     [Option("i", "id", Default = 1, Description = "Id of the user wallet to use")]
-    public uint Id { get; init; } = 1;
+    public int Id { get; init; } = 1;
 
     [CommandTarget]
     public async Task<int> Run()
@@ -16,7 +16,7 @@ internal sealed class DeleteUnconfirmedTransactionsCommand : WalletCommand
         return await DoWorkAsync("Deleting unconfirmed transactions...", async output =>
         {
             using var rpcClient = await ClientFactory.Factory.CreateRpcClient(output, this, ServiceNames.Wallet);
-            var wallet = new chia.dotnet.Wallet(Id, await Login(rpcClient, output));
+            var wallet = new chia.dotnet.Wallet((uint)Id, await Login(rpcClient, output));
 
             using var cts = new CancellationTokenSource(TimeoutMilliseconds);
             await wallet.DeleteUnconfirmedTransactions(cts.Token);
